requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
passlib==1.7.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
pywizlight==0.5.14
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.13
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
bcrypt==4.0.1
email-validator==2.0.0.post2
alembic==1.11.1
PyJWT==2.7.0
Pillow==9.5.0
Faker==18.9.0
psycopg2~=2.9.6
spotipy~=2.23.0
requests~=2.31.0